+++ src/App.test.jsx
@@ -10,6 +10,7 @@ jest.mock('react-redux');
 
 test('App', () => {
   useSelector.mockImplementation((selector) => selector({
+    taskTitle: '',
     tasks: [
       { id: 1, title: '할일 #1' },
       { id: 2, title: '할일 #2' },
@@ -19,6 +20,9 @@ test('App', () => {
   const { getByText } = render((
     <App />
   ));
+
+  expect(getByText(/할일 #1/)).not.toBeNull();
+  expect(getByText(/할일 #2/)).not.toBeNull();
 
   expect(getByText(/추가/)).not.toBeNull();
