+++ src/ListContainer.jsx
@@ -22,7 +22,7 @@ export default function ListContainer() {
   return (
     <List
       tasks={tasks}
-      onClickDeleteTask={handleClickDeleteTask}
+      onClickDelete={handleClickDeleteTask}
     />
   );
 }
